Made a new branch because bugFixer had a bug, how ironic. Will list changed made in comment
