+++ src/Home/Profile.js
@@ -121,14 +121,6 @@ class Profile extends PureComponent{
                         email: emailRes.data.email
                     })
                 }))
-                // axios.post('https://studybuddys-223920.appspot.com/api/getUsername', info)
-                // .then(response => {
-                //     page.setState({name : response.data.name})
-                // })
-                // axios.post('https://studybuddys-223920.appspot.com/api/getUseremail', info)
-                // .then(response => {
-                //     page.setState({email: response.data.email})
-                // })
             }
         });
     }
@@ -548,30 +540,29 @@ class Profile extends PureComponent{
                             this.state.deniedMeetup.map((title, key) => {
                             return(
                                 <Paper className="event" key = {key} style = {{width: "350px"}}>
-                                    <Typography variant = "title">
-                                        {this.state.deniedMeetup[key].courseName}
-                                    </Typography>
-
-                                    <Typography variant = "subtitle1">
-                                        Host Buddy: {this.state.deniedMeetup[key].name}
-                                    </Typography>
-
-                                    <Typography variant = "subtitle1">
-                                        Study Buddy: {this.state.deniedMeetup[key].partner}
-                                    </Typography>
-
-                                    <Typography variant = "subtitle1">
-                                        "{this.state.deniedMeetup[key].description}"
-                                    </Typography>
-
-                                    <Typography variant = "subtitle1">
-                                        {this.state.deniedMeetup[key].date}
-                                    </Typography>
-
-                                    <Typography variant = "subtitle1">
-                                        {this.state.deniedMeetup[key].time}
-                                    </Typography>
-
+                                     <Typography variant = "title">
+                                         {this.state.deniedMeetup[key].courseName}
+                                     </Typography>
+                                     
+                                     <Typography variant = "subtitle1">
+                                         Host Buddy: {this.state.deniedMeetup[key].name}
+                                     </Typography>
+                                     
+                                     <Typography variant = "subtitle1">
+                                         Study Buddy: {this.state.deniedMeetup[key].partner}
+                                     </Typography>
+                                     
+                                     <Typography variant = "subtitle1">
+                                         "{this.state.deniedMeetup[key].description}"
+                                     </Typography>
+                                     
+                                     <Typography variant = "subtitle1">
+                                         {this.state.deniedMeetup[key].date}
+                                     </Typography> 
+                                     
+                                     <Typography variant = "subtitle1">
+                                         {this.state.deniedMeetup[key].time}
+                                     </Typography>
                                     <Button onClick={()=>{this.setState({selectedMeetUp:this.state.deniedMeetup[key].meetupId},this.deleteMeetUps)}} color="secondary">Delete</Button>
                                 </Paper>
                             )
