+++ src/Home/page.css
@@ -1,0 +1,7 @@
+.pika {
+    background-image: url("./pika.png");
+    background-repeat: no-repeat;
+     background-size: 100% 100%;
+    background-attachment: relative;
+    height: 100vh;
+}
